Add deploy and sleep params to deploy plan
Useful features discovered in Nest. With YAML plans, it is nice to pass
a simple boolean parameter to enable or disable a deployment among a
list of deployments. YAML plans don't have conditional logic themselves.
Also, it is nice to let the deploy plan wait a little extra time if you
know it takes time for certain resources, like VIPs, to settle.